+++ conf.py
@@ -129,7 +129,7 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "venv","demonstrations/adjoint_timing.py"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "venv"]
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = "sphinx"
